+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # NEXT UP
 
 - Misty step effect not applied if using Automated Animations as it has better built in support.
+- Quasit effects for Claws.
+- Invisibility feature effect (not spell), for monsters.
 
 # 3.4.57